Backend/requirements.txt
django
djangorestframework
django-cors-headers
django-filter
djoser
python-dotenv
python-decouple
drf-spectacular
djangorestframework-simplejwt
gunicorn
whitenoise
psycopg2-binary
dj-database-url
